src/db_service/requirements.txt
psycopg2-binary==2.9.10
grpcio
grpcio-tools
protobuf
